backend/requirements.txt
# Web 框架
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# 数据库 ORM
sqlalchemy==2.0.35
alembic==1.13.3
psycopg[binary]==3.2.3
pgvector==0.3.4

# 数据校验
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# 鉴权（密码仅用 bcrypt 库；不再使用 passlib，避免与 bcrypt 4.x 冲突导致注册 500）
python-jose[cryptography]==3.3.0
bcrypt==4.2.0

# 缓存 / 队列
redis==5.1.0
celery==5.4.0

# 对象存储
boto3==1.35.36
minio==7.2.9

# 文件解析
python-docx==1.1.2
python-pptx==1.0.2
pypdf==5.0.1
pdfplumber==0.11.4
pytesseract==0.3.13
pdf2image==1.17.0
openpyxl==3.1.5
pandas==2.2.3
unstructured==0.15.14
markdownify==0.13.1
chardet==5.2.0

# AI / RAG
openai==1.51.0
tiktoken==0.8.0
numpy==1.26.4
scikit-learn==1.5.2

# HTTP 客户端
httpx==0.27.2
tenacity==9.0.0

# 工具
python-slugify==8.0.4
pyyaml==6.0.2
loguru==0.7.2

# 开发测试
pytest==8.3.3
pytest-asyncio==0.24.0
